Move common hardware agnostic logic from common_ipu.py into common.py
Many of these functions aren't necessarily tied to the ipu so moving it
to the common module makes more sense
